Combine curate_msa and create_unaligned_coding_seqs into single script
Combines the functionality of curate_msa.py and create_unaligned_coding_seqs.py
into a unified curate_and_extract_coding_seqs.py script for improved performance
and consistency.

Key improvements:
- Single-pass processing: Each sequence is filtered and validated once
- Stricter filtering: Sequences must pass ALL gene validations to be included
- Counter-based logging: Tracks statistics instead of logging every failure
- No intermediate I/O: Eliminates read/write overhead between steps
- Expected performance gain: 20-40% faster for large datasets

The combined approach ensures curated MSA contains only sequences with complete,
valid coding sequences for all genes. Log files are dramatically smaller and
more readable with summary statistics.

Tested successfully on HA/H7 (6,545 sequences processed in ~4 seconds).
